+++ core/src/main/java/cloud/stackit/sdk/core/utils/TestUtils.java
@@ -1,0 +1,6 @@
+package cloud.stackit.sdk.core.utils;
+
+public class TestUtils {
+	public static final String MOCK_SERVICE_ACCOUNT_KEY =
+			"{\"id\":\"id\",\"publicKey\":\"publicKey\",\"createdAt\":\"2025-03-26T15:08:45.915+00:00\",\"keyType\":\"keyType\",\"keyOrigin\":\"keyOrigin\",\"keyAlgorithm\":\"keyAlgo\",\"active\":true,\"validUntil\":\"2025-03-26T15:08:45.915+00:00\",\"credentials\":{\"aud\":\"aud\",\"iss\":\"iss\",\"kid\":\"kid\",\"privateKey\":\"privateKey\",\"sub\":\"sub\"}}\n";
+}